Fixes an issue where a worker's state could change to finished before the attached dialog was shown.
This was caused by the fix for #67.  Since the dialog is being shown after all the state listeners for a worker run, it's possible for one of those listeners to update the worker's state.  A case where a listener is used to immediately change the worker's state from SCHEDULED to CANCELLED is a good example.