#1114 Notifications: Fix notifications position
The position of the elements does not use the notification position
anymore. It uses the position where the notifications will be shown.

The following errors are fixed with this change:
- Fix old notifications shown not aligned with new notifications
when the window is moved
- Fix animation, it does not use the position of the notification
because if the animation is not finished, the position is not
correct. Previously, the position between the first and second
notification was not right if there were several notifications.