Allow user to overwrite the rating icon
As of now it not possible to easily customize the look of the rating control apart from changing the look via css. Other customization like using font-based glyphs for the star are not possible.

This PR makes the creation of the star node protected so that one customize it by overriding.