Fixes #1606. Existing behaviour never worked as expected, nodes were removed from expandedNodeCache (expandedNodeCache.remove(getBean())) everytime a row was collapsed.
This meant that ExpandableTableRowSkin's tableRow.itemProperty() listener was was getting null from the cache, and unable to remove the children, to clean-up the row.

This problem went unnoticed until now because a table refresh was re-creating all cells anyway. Now though, there are optimisations in TableView which means these leftover nodes are visible / never cleaned up.